Update `nodejs-corepack` to use new layer API
